Enhance e-commerce backend functionality and improve code structure
- Introduced new features for order management, including order creation, shipping, and refund processing.
- Added support for batch operations in the cart service, allowing users to update quantities and manage items more efficiently.
- Refactored user and merchant management, including enhanced security annotations and role-based access control.
- Updated database schema with new columns for order items and added migration scripts for better data handling.
- Improved API documentation and logging for better traceability and user experience.
- Cleaned up and organized project structure, removing obsolete files and enhancing overall maintainability.

These changes significantly enhance the functionality and organization of the SuperMall e-commerce backend.

+++ src/main/java/com/supermall/backend/domain/cart/entity/CartItem.java
@@ -8,9 +8,9 @@ import java.time.LocalDateTime;
 @TableName("shopping_cart_items")
 public class CartItem {
     @TableId(value = "item_id", type = IdType.AUTO)
-    private Long id;
-    private Long userId;
-    private Long productId;
+    private Integer id;
+    private Integer userId;
+    private Integer productId;
     private Integer quantity;
     private Boolean selected;
