+++ src/main/java/com/supermall/backend/domain/user/service/UserAddressService.java
@@ -5,11 +5,11 @@ import com.supermall.backend.domain.user.dto.AddressResponse;
 import java.util.List;
 
 public interface UserAddressService {
-    AddressResponse createAddress(Long userId, AddressRequest request);
-    AddressResponse updateAddress(Long userId, Long addressId, AddressRequest request);
-    void deleteAddress(Long userId, Long addressId);
-    AddressResponse getAddress(Long userId, Long addressId);
-    List<AddressResponse> listAddresses(Long userId);
-    void setDefaultAddress(Long userId, Long addressId);
-    AddressResponse getDefaultAddress(Long userId);
+    AddressResponse createAddress(Integer userId, AddressRequest request);
+    AddressResponse updateAddress(Integer userId, Integer addressId, AddressRequest request);
+    void deleteAddress(Integer userId, Integer addressId);
+    AddressResponse getAddress(Integer userId, Integer addressId);
+    List<AddressResponse> listAddresses(Integer userId);
+    void setDefaultAddress(Integer userId, Integer addressId);
+    AddressResponse getDefaultAddress(Integer userId);
 }
